Corrected sql file

+++ urlshortenerservice/conf/evolutions/default/1.sql
@@ -8,4 +8,4 @@ create table urls (
 
 # --- !Downs
 
-drop table urls;
+drop table urls if exists;
